AstroQuant_Phase1/requirements.txt
fastapi
uvicorn
databento
pandas
python-dotenv
requests
